refactor: 내 방 피드 데이터 추가 및 리팩토링
