+++ Module-API/src/main/java/depromeet/api/domain/feed/dto/ChallengeFeed.java
@@ -1,18 +1,9 @@
 package depromeet.api.domain.feed.dto;
 
 
-import com.fasterxml.jackson.annotation.JsonInclude;
-import depromeet.domain.comment.domain.Comment;
-import depromeet.domain.emoji.domain.Emoji;
-import depromeet.domain.record.domain.EmojiType;
 import depromeet.domain.record.domain.Record;
 import depromeet.domain.userchallenge.domain.UserChallenge;
 import io.swagger.v3.oas.annotations.media.Schema;
-import java.time.LocalDateTime;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-import java.util.stream.Collectors;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -46,6 +37,7 @@ public class ChallengeFeed {
                 .build();
     }
 
+    @Data
     public static class UserInfo {
         @Schema(example = "유저 이미지 URL")
         private String userImgUrl;
@@ -60,76 +52,6 @@ public class ChallengeFeed {
             this.userImgUrl = userChallenge.getImgUrl();
             this.nickname = userChallenge.getNickname();
             this.currentCharge = userChallenge.getCurrentCharge();
-        }
-    }
-
-    private static class RecordInfo {
-        @Schema(description = "지출 기록 ID", example = "27")
-        private Long recordId;
-
-        @Schema(example = "기록 이미지 URL")
-        private String recordImgUrl;
-
-        @Schema(example = "기록 타이틀")
-        private String title;
-
-        @Schema(example = "기록 내용")
-        private String content;
-
-        @Schema(description = "가격", example = "5000")
-        private Integer price;
-
-        @Schema(description = "기록 작성일")
-        private LocalDateTime recordDate;
-
-        public RecordInfo(Record record) {
-            this.recordId = record.getId();
-            this.recordImgUrl = record.getImgUrl();
-            this.title = record.getTitle();
-            this.content = record.getContent();
-            this.price = record.getPrice();
-            this.recordDate = record.getCreatedAt();
-        }
-    }
-
-    private static class EmojiInfo {
-
-        @JsonInclude(JsonInclude.Include.NON_NULL)
-        private String selectedEmoji;
-
-        @Schema(description = "미친거지", example = "2")
-        private Long crazy;
-
-        @Schema(description = "후회할거지", example = "0")
-        private Long regretful;
-
-        @Schema(description = "잘할거지", example = "3")
-        private Long wellDone;
-
-        @Schema(description = "댓글수", example = "5")
-        private Integer comment;
-
-        public EmojiInfo(Record record, Long userChallengeId) {
-            Optional<Emoji> emoji =
-                    record.getEmojis().stream()
-                            .filter((e) -> e.getUserChallenge().getId() == userChallengeId)
-                            .findAny();
-
-            if (emoji.isPresent()) this.selectedEmoji = emoji.get().getType().getValue();
-            else this.selectedEmoji = null;
-
-            Map<String, Long> emojiCountMap =
-                    record.getEmojis().stream()
-                            .map(Emoji::getType)
-                            .collect(
-                                    Collectors.groupingBy(
-                                            EmojiType::getValue, Collectors.counting()));
-            this.crazy = emojiCountMap.get(EmojiType.CRAZY_BEGGAR.getValue());
-            this.regretful = emojiCountMap.get(EmojiType.REGRETFUL_BEGGAR.getValue());
-            this.wellDone = emojiCountMap.get(EmojiType.WELL_DONE_BEGGAR.getValue());
-
-            List<Comment> comments = record.getComments();
-            this.comment = comments.size();
         }
     }
 }
