+++ Module-API/src/main/java/depromeet/api/domain/feed/dto/MyFeed.java
@@ -1,9 +1,9 @@
 package depromeet.api.domain.feed.dto;
 
 
+import depromeet.domain.challenge.domain.Challenge;
 import depromeet.domain.record.domain.Record;
 import io.swagger.v3.oas.annotations.media.Schema;
-import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -13,40 +13,35 @@ import lombok.Data;
 @Data
 public class MyFeed {
 
-    @Schema(description = "기록 Id", example = "27")
-    private Long recordId;
+    private RecordInfo recordInfo;
 
-    @Schema(example = "기록 이미지 URL")
-    private String recordImgUrl;
+    private ChallengeInfo challengeInfo;
 
-    @Schema(description = "가격", example = "5000")
-    private Integer price;
-
-    @Schema(example = "기록 타이틀")
-    private String title;
-
-    @Schema(example = "기록 내용")
-    private String content;
-
-    @Schema(example = "챌린지 이미지 URL")
-    private String challengeImgUrl;
-
-    @Schema(example = "챌린지 타이틀")
-    private String challengeTitle;
-
-    @Schema(description = "기록 작성일")
-    private LocalDateTime recordDate;
+    private EmojiInfo emojiInfo;
 
     public static MyFeed createMyFeed(Record record) {
+        RecordInfo recordInfo = new RecordInfo(record);
+        ChallengeInfo challengeInfo = new ChallengeInfo(record.getChallenge());
+        EmojiInfo emojiInfo = new EmojiInfo(record);
+
         return MyFeed.builder()
-                .recordId(record.getId())
-                .recordImgUrl(record.getImgUrl())
-                .price(record.getPrice())
-                .title(record.getTitle())
-                .content(record.getContent())
-                .challengeImgUrl(record.getChallenge().getImgUrl())
-                .challengeTitle(record.getChallenge().getTitle())
-                .recordDate(record.getCreatedAt())
+                .recordInfo(recordInfo)
+                .challengeInfo(challengeInfo)
+                .emojiInfo(emojiInfo)
                 .build();
+    }
+
+    @Data
+    public static class ChallengeInfo {
+        @Schema(example = "챌린지 이미지 URL")
+        private String imgUrl;
+
+        @Schema(example = "챌린지 타이틀")
+        private String title;
+
+        public ChallengeInfo(Challenge challenge) {
+            this.imgUrl = challenge.getImgUrl();
+            this.title = challenge.getTitle();
+        }
     }
 }
